Refactor: separate ES client into multiple files per func type
